# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "\u5bb6\u6709\u6cd5\u5b9d\u63a5\u53e3\u6587\u6863",
        "description": "\u5bb6\u6709\u6cd5\u5b9d\u662f\u4e00\u4e2a\u4f7f\u7528\u5ea6\u975e\u5e38\u9ad8\u7684\u5fae\u4fe1\u5c0f\u7a0b\u5e8f\uff0c\u53d1\u5e03\u81f3\u4eca\u83b7\u5f97\u4e86\u65e0\u6570\u7684\u7528\u6237\u597d\u8bc4\uff0c\u4e3a\u4e86\u6ee1\u8db3\u4f7f\u7528\u8005\u7684\u9700\u8981\uff0c\u73b0\u5c06\u63a5\u53e3\u5217\u8868\u8fd9\u91cc\u5982\u4e0b\uff0c\u5982\u6709\u95ee\u9898\u6b22\u8fce\u53ca\u65f6\u53cd\u9988<br>\r\n\u4ee5\u540e\u5927\u5bb6\u5c31\u5728\u8fd9\u91cc\u6109\u5feb\u7684\u5bf9\u63a5\u53e3\u628a!<br>\r\n\u4ee5\u540e\u5927\u5bb6\u5c31\u5728\u8fd9\u91cc\u6109\u5feb\u7684\u5bf9\u63a5\u53e3\u628a!<br>\r\n\u4ee5\u540e\u5927\u5bb6\u5c31\u5728\u8fd9\u91cc\u6109\u5feb\u7684\u5bf9\u63a5\u53e3\u628a!<br>",
        "contact": {
            "name": "joey",
            "email": "zongjun.li@outlook.com"
        },
        "version": "1.0"
    },
    "host": "dev.laravel.com/api/v1",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\u7528\u6237\u8d44\u6599\u7ba1\u7406",
                "description": "\u7528\u6237\u8d44\u6599\u5217\u8868",
                "operationId": "user.index",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "\u5206\u9875\u7f16\u53f7,\u9ed8\u8ba41",
                        "type": "integer"
                    },
                    {
                        "name": "pagesize",
                        "in": "query",
                        "description": "\u6bcf\u9875\u663e\u793a\u6761\u6570,\u9ed8\u8ba410",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u7528\u6237\u5217\u8868",
                        "schema": {
                            "properties": {
                                "pages": {
                                    "properties": {
                                        "totalnum": {
                                            "description": "\u8bb0\u5f55\u603b\u6570",
                                            "type": "integer"
                                        },
                                        "totalpage": {
                                            "description": "\u603b\u9875\u6570",
                                            "type": "integer"
                                        },
                                        "pagesize": {
                                            "description": "\u6bcf\u9875\u663e\u793a\u8bb0\u5f55\u6570",
                                            "type": "integer"
                                        },
                                        "page": {
                                            "description": "\u5f53\u524d\u9875\u6570",
                                            "type": "integer"
                                        }
                                    }
                                },
                                "lists": {
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "id",
                                                "type": "integer"
                                            },
                                            "username": {
                                                "description": "\u7528\u6237\u540d",
                                                "type": "string"
                                            },
                                            "nickname": {
                                                "description": "\u6635\u79f0",
                                                "type": "string"
                                            },
                                            "avatar": {
                                                "description": "\u5934\u50cf",
                                                "type": "string"
                                            },
                                            "mobile": {
                                                "description": "\u624b\u673a\u53f7\u7801",
                                                "type": "string"
                                            },
                                            "descr": {
                                                "description": "\u63cf\u8ff0",
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            },
                            "type": "json"
                        }
                    },
                    "422": {
                        "description": "error"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "\u7528\u6237\u6ce8\u518c",
                "description": "\u7528\u6237\u6ce8\u518c\u63a5\u53e3",
                "parameters": [
                    {
                        "name": "username",
                        "in": "formData",
                        "description": "\u7528\u6237\u540d",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "\u767b\u5f55\u5bc6\u7801",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u6210\u529f"
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\u7528\u6237\u8be6\u60c5",
                "description": "\u7528\u6237\u8be6\u60c5\u63a5\u53e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "\u7528\u6237id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u6210\u529f"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "\u767b\u5f55\u63a5\u53e3(\u7528\u6237\u540d+\u5bc6\u7801)",
                "description": "\u7528\u6237\u767b\u5f55\u63a5\u53e3,\u8d26\u53f7\u53ef\u4e3a \u7528\u6237\u540d \u6216 \u624b\u673a\u53f7",
                "parameters": [
                    {
                        "name": "username",
                        "in": "formData",
                        "description": "\u767b\u5f55\u7528\u6237\u540d/\u624b\u673a\u53f7",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "\u767b\u5f55\u5bc6\u7801",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "client_type",
                        "in": "formData",
                        "description": "\u8c03\u7528\u6b64\u63a5\u53e3\u7684\u5ba2\u6237\u7aef\u7c7b\u578b: 1-Android, 2-IOS. \u975e\u5fc5\u586b,\u6240\u4ee5 required \u5199\u4e86 false",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u6210\u529f"
                    }
                }
            }
        },
...
        "/users/info": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\u4e2a\u4eba\u4fe1\u606f",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "\u6210\u529f"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/users/comments/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\u6211\u7684\u8bc4\u8bba",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "\u7528\u6237id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u6210\u529f"
                    },
                    "403": {
                        "description": "\u7981\u6b62\u8bbf\u95ee"
                    },
                    "404": {
                        "description": "\u672a\u627e\u5230"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/users/collections/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\u6211\u7684\u6536\u85cf",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "\u7528\u6237id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u6210\u529f"
                    },
                    "403": {
                        "description": "\u7981\u6b62\u8bbf\u95ee"
                    },
                    "404": {
                        "description": "\u672a\u627e\u5230"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/users/browse-historys/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\u6211\u7684\u6536\u85cf",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "\u7528\u6237id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u6210\u529f"
                    },
                    "403": {
                        "description": "\u7981\u6b62\u8bbf\u95ee"
                    },
                    "404": {
                        "description": "\u672a\u627e\u5230"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/users/balance-logs/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\u6211\u7684\u6536\u85cf",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "\u7528\u6237id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u6210\u529f"
                    },
                    "403": {
                        "description": "\u7981\u6b62\u8bbf\u95ee"
                    },
                    "404": {
                        "description": "\u672a\u627e\u5230"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}